Replaced message store with message reaction
