+++ app/scripts/components/channel.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import connectToStores from 'alt/utils/connectToStores';
 
-import MessageStore from '../stores/messages';
 import ServerStore from '../stores/servers';
 
 import UserList from './userlist';
@@ -10,10 +9,9 @@ import Input from './input';
 
 @connectToStores
 export default class ChannelView extends React.Component {
-    static getStores() { return [MessageStore, ServerStore]; }
+    static getStores() { return [ServerStore]; }
     static getPropsFromStores() {
         return {
-            messages: MessageStore.getState().messages,
             servers: ServerStore.getState().servers
         };
     }
@@ -25,8 +23,8 @@ export default class ChannelView extends React.Component {
         const oldChannelName = decodeURIComponent(this.props.params.channel);
         const newChannelName = decodeURIComponent(newProps.params.channel);
 
-        const oldMessages = this.props.messages[oldServerId].channels[oldChannelName];
-        const newMessages = newProps.messages[newServerId].channels[newChannelName];
+        const oldMessages = this.props.state.messages[oldServerId].channels[oldChannelName];
+        const newMessages = newProps.state.messages[newServerId].channels[newChannelName];
 
         const oldChannel = this.props.servers[oldServerId].channels[oldChannelName];
         const newChannel = newProps.servers[newServerId].channels[newChannelName];
@@ -36,7 +34,7 @@ export default class ChannelView extends React.Component {
 
     render() {
         const { serverId } = this.props.params;
-        const { messages, servers } = this.props;
+        const { messages, servers } = this.props.state;
         const channel = decodeURIComponent(this.props.params.channel);
 
         let chanMessages = null;
